feat: create cursor pagination util
Offset pagination isn't scalable for two reasons:
- retrieving count (full scan without caching strategy)
- offsetting results (scan first pages to reach wanted page)

Cursor pagination solves that issue by using cursors to directly
point to the location of the previous results limits.

Internally, this is transformed into basic comparisons,
which is really fast with indexes.

The drawback is that count isn't available anymore
and there is no way to directly jump into a specific page.